Consolidating page-endpoints
There were too many files so one file for all endpoints delivering pages
There's still some little mixup in that the logs work as both api and
page server (to be fixed).

Adding a status api for things that was previously entangled with page
endpoint
