--- scanomatic/ui_server/help_ui.py
@@ -1,20 +1,0 @@
-from flask import send_from_directory, redirect
-
-from scanomatic.io.paths import Paths
-
-
-def add_routes(app):
-    """
-
-    :param app: The flask webb app
-     :type app: Flask
-    :return:
-    """
-
-    @app.route("/help")
-    def _help():
-        return send_from_directory(Paths().ui_root, Paths().ui_help_file)
-
-    @app.route("/wiki")
-    def _wiki():
-        return redirect("https://github.com/local-minimum/scanomatic/wiki")
